fix: resolve compilation errors in GameSession and EntityFactory
- Fix ModuleLoader namespace conflict by using fully qualified Loading.ModuleLoader
- Fix IEntity.World to be settable (change from get-only to get/set)
- Fix EntityFactory.GetObjectIdFromGit to cast int to uint for ObjectId
- Remove invalid entity.World assignment (World is set during entity creation)
- All compilation errors resolved, build succeeds

+++ src/OdysseyRuntime/Odyssey.Kotor/Loading/EntityFactory.cs
@@ -80,7 +80,7 @@ namespace Odyssey.Kotor.Loading
         private uint GetObjectIdFromGit(GFFStruct gitStruct)
         {
             // Try to read ObjectId from GIT (may not always be present)
-            uint objectId = GetIntField(gitStruct, "ObjectId", 0);
+            uint objectId = (uint)GetIntField(gitStruct, "ObjectId", 0);
             if (objectId != 0 && objectId != 0x7F000000) // 0x7F000000 = OBJECT_INVALID
             {
                 return objectId;
